// frontend/src/types/api/actionType.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * mailassist
 * OpenAPI spec version: 0.1.0
 */

/**
 * Available rule action types.
 */
export type ActionType = typeof ActionType[keyof typeof ActionType];


export const ActionType = {
  move: 'move',
  copy: 'copy',
  label: 'label',
  remove_label: 'remove_label',
  mark_read: 'mark_read',
  mark_unread: 'mark_unread',
  flag: 'flag',
  delete: 'delete',
  notify: 'notify',
  create_draft: 'create_draft',
  create_calendar_event: 'create_calendar_event',
} as const;